Add ability to update the common_name of SmartCsr
In order to authenticate hosts from every policy branch, we need the
authn-client to send the full host-id to the server. To maintain backwards
compatibility we need to support also client versions that do not send the
full host-id, and assume that the server adds the host-id prefix to it.
This commit adds the ability to update the CSR so the server can do so
before the CSR validations, to make both scenarios as similar as possible