Update the CSR's common-name in case of an old client
In order to authenticate hosts from every policy branch, we need the
authn-client to send the full host-id to the server. To maintain backwards
compatibility we need to support also client versions that do not send the
full host-id, and assume that the server adds the host-id prefix to it.

This commit updates the CSR in case of an old client so it will include
also the hard-coded prefix of the "apps" policy branch