Add UT for an empty Host-Id-Prefix parameter
In case the inject_client_cert request is called without the Host-Id-Prefix
parameter in the header we need to update the common-name and the injection
shouldn't fail. This UT verifies this.

Also, we fix the rest of the UTs to send the Host-Id-Prefix parameter as
the behavior of inject_client_cert has changed