Revert to using master branch errors.rb
During the rebase on Oren's authenticator work, the formatting changes
and the "||=" that had been made to this file were causing errors.

Eventually I just reverted to the code to what was in master, because
fixing them was becoming a rabbit hole.

This cleanup of the errors.rb file is a separate task from the Rails
refactor, and can be revisited later.